Exclude unfetchable dependencies
Pentaho dependencies were unfetchable because the conjars repo
changed of URL. This was fixed in HIVE-25173, I'm applying more
or less the same action.

For the others, some errors popped up about the repo being 'blocked',
looking online it could be related to them being in http rather than
https.